Pass floating point numbers as qreal's to qMin/qMax, reported by krazy.
svn path=/trunk/KDE/kdeedu/marble/; revision=1214616

+++ src/lib/TileScalingTextureMapper.cpp
@@ -74,7 +74,8 @@ void TileScalingTextureMapper::mapTexture( ViewParams *viewParams )
     const int maxTileX = numTilesX * ( xNormalizedCenter + imageWidth/( 8.0 * radius ) );
 
     const qreal yNormalizedCenter = 0.5 - 0.5 * asinh( tan( centerLat ) ) / M_PI;
-    const int minTileY = qMax( numTilesY * ( yNormalizedCenter - imageHeight/( 8.0 * radius ) ), 0.0 );
+    const int minTileY = qMax( numTilesY * ( yNormalizedCenter - imageHeight/( 8.0 * radius ) ),
+                               qreal( 0.0 ));
     const int maxTileY = qMin( numTilesY * ( yNormalizedCenter + imageHeight/( 8.0 * radius ) ), numTilesY - 1.0 );
 
     QPainter painter( canvasImage );
